Scripts/requirements.txt
mysql-connector-python
python-dotenv
selenium
webdriver_manager